Add documentation and refactor - Add package-level documentation and method-level documentation - Refactor larger functions
