+++ api/impl.go
@@ -1,19 +1,11 @@
 package api
 
 import (
-	"context"
-	"database/sql"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"net/http"
-	"strings"
-	"time"
 
-	"github.com/SlotifyApp/slotify-backend/database"
-	"github.com/SlotifyApp/slotify-backend/jwt"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 const (
@@ -47,79 +39,6 @@ func SetHeaderAndWriteResponse(w http.ResponseWriter, code int, encode any) {
 	}
 }
 
-type (
-	Users []User
-	Teams []Team
-)
-
-// MarshalLogArray implements the zapcore.ArrayMarshaler interface.
-func (u Users) MarshalLogArray(arr zapcore.ArrayEncoder) error {
-	for _, user := range u {
-		if err := arr.AppendObject(user); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler.
-
-func (u UserCreate) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	enc.AddString("email", string(u.Email))
-	enc.AddString("firstName", u.FirstName)
-	enc.AddString("lastName", u.LastName)
-	return nil
-}
-
-func (u User) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	userCreate := UserCreate{
-		Email:     u.Email,
-		FirstName: u.FirstName,
-		LastName:  u.LastName,
-	}
-	if err := userCreate.MarshalLogObject(enc); err != nil {
-		return fmt.Errorf("failed to marshal User obj: %v", err.Error())
-	}
-	enc.AddUint32("id", u.Id)
-	return nil
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler.
-func (t TeamCreate) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	enc.AddString("name", t.Name)
-	return nil
-}
-
-// MarshalLogArray implements the zapcore.ArrayMarshaler interface.
-func (t Teams) MarshalLogArray(arr zapcore.ArrayEncoder) error {
-	for _, team := range t {
-		if err := arr.AppendObject(team); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (t Team) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	teamCreate := TeamCreate{
-		Name: t.Name,
-	}
-	if err := teamCreate.MarshalLogObject(enc); err != nil {
-		return fmt.Errorf("failed to marshal Team obj: %v", err.Error())
-	}
-	enc.AddUint32("id", t.Id)
-	return nil
-}
-
-func (tp GetAPITeamsParams) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	name := ""
-	if tp.Name != nil {
-		name = *tp.Name
-	}
-	enc.AddString("name", name)
-	return nil
-}
-
 // (GET /healthcheck).
 func (s Server) GetAPIHealthcheck(w http.ResponseWriter, _ *http.Request) {
 	resp := "Healthcheck Successful!"
@@ -129,184 +48,4 @@ func (s Server) GetAPIHealthcheck(w http.ResponseWriter, _ *http.Request) {
 		sendError(w, http.StatusInternalServerError, "Failed to encode JSON")
 		return
 	}
-}
-
-// splitName splits a name into first name and last name.
-func splitName(name string) (string, string) {
-	names := strings.Fields(name)
-	var firstName string
-	var lastName string
-
-	if len(names) > 0 {
-		firstName = names[0]
-	}
-
-	if len(names) > 1 {
-		// Join the rest of the fields together to form the
-		// last name
-		lastName = strings.Join(names[1:], " ")
-	}
-	return firstName, lastName
-}
-
-type MSFTEntraValues struct {
-	ClientID     string
-	ClientSecret string
-	TenantID     string
-}
-
-// getOrInsertUserByClaimEmail will get user by the claim email or if first time log in, it will create a new user.
-func getOrInsertUserByClaimEmail(qtx *database.Queries, msftTokenRes MSFTTokenResult) (database.User, error) {
-	email := msftTokenRes.Email
-	// Double the timeout due to more db operations
-	ctx, cancel := context.WithTimeout(context.TODO(), 20*time.Second)
-	defer cancel()
-
-	count, err := qtx.CountUserByEmail(ctx, email)
-	if err != nil {
-		return database.User{}, fmt.Errorf("failed to get user count by claim email: %w", err)
-	}
-
-	// User doesn't exist, first time signing so create a new user
-	if count == 0 {
-		dbParams := database.CreateUserParams{
-			Email:     email,
-			FirstName: msftTokenRes.FirstName,
-			LastName:  msftTokenRes.LastName,
-		}
-		_, err = qtx.CreateUser(ctx, dbParams)
-		if err != nil {
-			return database.User{}, fmt.Errorf("failed to create user for claim email: %w", err)
-		}
-	}
-
-	var u database.User
-
-	u, err = qtx.GetUserByEmail(ctx, email)
-	if err != nil {
-		return database.User{}, fmt.Errorf("failed to get user with claim email: %w", err)
-	}
-
-	// Update user's home account id so it can be used when asking for a MSFT access token
-	dbParams := database.UpdateUserHomeAccountIDParams{
-		ID:                u.ID,
-		MsftHomeAccountID: sql.NullString{String: msftTokenRes.HomeAccountID, Valid: true},
-	}
-	var rowsAffected int64
-	rowsAffected, err = qtx.UpdateUserHomeAccountID(ctx, dbParams)
-	if err != nil {
-		return database.User{}, fmt.Errorf("failed to update user home account id: %w", err)
-	}
-
-	// UpdateUserHomeAccountID will either update 1 or 0 rows.
-	if rowsAffected > 1 {
-		err = database.WrongNumberSQLRowsError{
-			ActualRows:   rowsAffected,
-			ExpectedRows: []int64{0, 1},
-		}
-		return database.User{}, fmt.Errorf("failed to update home account id: %w", err)
-	}
-
-	return u, nil
-}
-
-func (s Server) GetAPIAuthCallback(w http.ResponseWriter, r *http.Request, params GetAPIAuthCallbackParams) {
-	msftTokenRes, err := msftAuthoriseByCode(r.Context(), s.MSALClient, params.Code)
-	if err != nil {
-		s.Logger.Error("failed to get microsoft tokens", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, "Sorry, try again later. Failed to get Microsoft tokens.")
-		return
-	}
-
-	tx, err := s.DB.DB.Begin()
-	if err != nil {
-		s.Logger.Error("failed to start db transaction", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, "callback route: failed to start db transaction")
-		return
-	}
-
-	defer func() {
-		if err = tx.Rollback(); err != nil {
-			s.Logger.Error("failed to rollback db transaction", zap.Error(err))
-		}
-	}()
-
-	qtx := s.DB.WithTx(tx)
-	var u database.User
-	if u, err = getOrInsertUserByClaimEmail(qtx, msftTokenRes); err != nil {
-		s.Logger.Error("failed to get user for claim email from msft access token", zap.Error(err))
-		sendError(w, http.StatusBadRequest, "failed to parse msft access token")
-		return
-	}
-
-	var tks jwt.AccessAndRefreshTokens
-	if tks, err = jwt.CreateAccessAndRefreshTokens(r.Context(), s.Logger, qtx, u.ID, u.Email); err != nil {
-		s.Logger.Error("failed to create and store tokens", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, "failed to create slotify access and refresh token")
-		return
-	}
-
-	if err = tx.Commit(); err != nil {
-		s.Logger.Error("failed to commit db transaction", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, "failed to commit db transaction")
-		return
-	}
-
-	CreateCookies(w, tks.AccessToken, tks.RefreshToken)
-
-	http.Redirect(w, r, "http://localhost:3000/dashboard", http.StatusFound)
-}
-
-func (s Server) PostAPIRefresh(w http.ResponseWriter, r *http.Request) {
-	ctx, cancel := context.WithTimeout(r.Context(), 5*database.DatabaseTimeout)
-	defer cancel()
-
-	var refreshToken string
-	if refreshToken = r.Header.Get(refreshTokenHeader); refreshToken == "" {
-		s.Logger.Error("refresh token was empty")
-		sendError(w, http.StatusUnauthorized, "refresh token was empty")
-		return
-	}
-	claims, err := jwt.ParseJWT(refreshToken, jwt.RefreshTokenJWTSecretEnv)
-	if err != nil {
-		s.Logger.Errorf("failed to verify refreshToken", zap.Error(err))
-		sendError(w, http.StatusUnauthorized, "refresh token was invalid")
-		return
-	}
-
-	userID := claims.UserID
-
-	var rt database.RefreshToken
-	if rt, err = s.DB.GetRefreshTokenByUserID(ctx, userID); err != nil {
-		s.Logger.Error("failed to get refresh token for user", zap.Error(err))
-		sendError(w, http.StatusUnauthorized, "failed to refresh token")
-		return
-	}
-
-	// check if the actual user's refresh token matches the request's refresh token
-	if rt.Token != refreshToken || rt.Revoked {
-		s.Logger.Error("Failed to match provided token or verify token OR token was revoked", zap.Error(err))
-		sendError(w, http.StatusUnauthorized, "failed to refresh token")
-		return
-	}
-
-	// Generate new access token and new refresh token
-	var uq database.User
-	if uq, err = s.DB.GetUserByID(ctx, userID); err != nil {
-		s.Logger.Error("Failed to refresh token", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	var tks jwt.AccessAndRefreshTokens
-	tks, err = jwt.CreateAccessAndRefreshTokens(ctx, s.Logger, &s.DB.Queries, userID, uq.Email)
-	if err != nil {
-		s.Logger.Error("failed to create access and refresh tokens", zap.Error(err))
-		sendError(w, http.StatusInternalServerError, "failed to create access and refresh tokens")
-		return
-	}
-
-	CreateCookies(w, tks.AccessToken, tks.RefreshToken)
-
-	SetHeaderAndWriteResponse(w, http.StatusCreated, "Successfully refreshed tokens")
 }
