+++ api_docs/.openapi-generator/FILES
@@ -1,3 +1,4 @@
+.openapi-generator-ignore
 Apis/DefaultApi.md
 Models/Attendee.md
 Models/CalendarEvent.md
