+++ integration/team_impl_test.go
@@ -140,7 +140,7 @@ func TestTeam_PostTeams(t *testing.T) {
 			req := httptest.NewRequest(http.MethodPost, "/api/teams", bytes.NewReader(body))
 			req.Header.Add("Content-Type", "application/json")
 
-			ctx := context.WithValue(req.Context(), api.UserCtxKey{}, user.Id)
+			ctx := context.WithValue(req.Context(), api.UserIDCtxKey{}, user.Id)
 			req = req.WithContext(ctx)
 
 			server.PostAPITeams(rr, req)
@@ -484,7 +484,7 @@ func TestTeam_GetAPITeamsMe(t *testing.T) {
 			rr := httptest.NewRecorder()
 			req := httptest.NewRequest(http.MethodGet, "/api/teams/me", nil)
 
-			ctx := context.WithValue(req.Context(), api.UserCtxKey{}, tt.userID)
+			ctx := context.WithValue(req.Context(), api.UserIDCtxKey{}, tt.userID)
 			req = req.WithContext(ctx)
 
 			server.GetAPITeamsMe(rr, req)
