+++ logger/logger.go
@@ -6,11 +6,12 @@ import (
 	"go.uber.org/zap"
 )
 
-// Logger just wraps a zap.SugaredLogger.
+// Logger creates a logger that wraps a zap.SugaredLogger.
 type Logger struct {
 	*zap.SugaredLogger
 }
 
+// NewLogger returns a new instance of Logger.
 func NewLogger() (*Logger, error) {
 	logger, err := zap.NewProduction()
 	if err != nil {
