+++ notification/errors.go
@@ -4,4 +4,5 @@ import (
 	"errors"
 )
 
+// ErrNotifClientNil is returned if the notification service attempts to register a nil client.
 var ErrNotifClientNil = errors.New("notification client for user cannot be nil")
